requirements_dev.txt
pip==23.3.2 
wheel==0.42.0
twine==4.0.2
black==23.12.1
python-dotenv==1.0.0
pytest==7.4.4
pytest-cov==4.1.0
sphinx==7.2.6
sphinx-autoapi==3.0.0
